@@ -1,33 +1,23 @@
 {
-  "name": "manna-desktop",
-  "version": "0.1.0",
-  "private": true,
+  "name": "manna-desktop-v3",
+  "version": "0.0.0",
   "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "test:e2e": "vue-cli-service test:e2e"
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview --port 4173"
   },
   "dependencies": {
-    "@folder/xdg": "^2.1.1",
-    "@mdi/font": "^5.9.55",
-    "buefy": "^0.8.20",
-    "iso-639-1": "^2.1.8",
-    "lodash": "^4.17.21",
-    "scrollbooster": "^2.3.0",
-    "thenby": "^1.3.4",
-    "vue": "^2.6.12",
-    "vue-context": "^5.2.0",
-    "vue-router": "^3.5.1",
-    "vuex": "^3.6.2"
+    "pinia": "^2.0.17",
+    "vue": "^3.2.37",
+    "vue-router": "^4.1.3"
   },
   "devDependencies": {
-    "@vue/cli-plugin-e2e-cypress": "~4.3.0",
-    "@vue/cli-plugin-router": "~4.3.0",
-    "@vue/cli-plugin-vuex": "~4.3.0",
-    "@vue/cli-service": "~4.3.0",
-    "electron": "^8.5.5",
-    "less": "^3.13.1",
-    "less-loader": "^5.0.0",
-    "vue-template-compiler": "^2.6.12"
+    "@vitejs/plugin-vue": "^3.0.1",
+    "autoprefixer": "^10.4.8",
+    "less": "^4.1.3",
+    "less-loader": "^11.0.0",
+    "postcss": "^8.4.16",
+    "tailwindcss": "^3.1.8",
+    "vite": "^3.0.4"
   }
 }
